requirements.txt
setuptools
Pillow>=3.2.0
